+++ diive/__init__.py
@@ -1,4 +1,15 @@
+from diive.core.dfun.frames import transform_yearmonth_matrix_to_longform as transform_yearmonth_matrix_to_longform
+
+# Plotting
+from diive.core.plotting.heatmap_datetime import HeatmapDateTime as heatmapdatetime
+from diive.core.plotting.heatmap_datetime import HeatmapYearMonth as heatmapyearmonth
+from diive.core.plotting.heatmap_datetime import HeatmapYearMonthRanks as heatmapyearmonth_ranks
 from diive.core.plotting.ridgeline import RidgeLinePlot as ridgeline
+
+from diive.core.times.resampling import resample_to_monthly_agg_matrix as resample_to_monthly_agg_matrix
+from diive.pkgs.createvar.conversions import et_from_le as et_from_le
+
+
 
 # import diive.core.plotting as plot
 # plot.ridgeline()